Add front correctives for datatable

+++ resources/js/components/Users.vue
@@ -1,99 +1,111 @@
 <template>
-    <div class="container" v-if="users">
-        <button type="button" @click.prevent="open('add')" data-bs-toggle="modal" data-bs-target="#staticBackdrop"
-                class="btn btn-primary mb-5">Add
-            user
-        </button>
-        <table class="table" v-if="users">
-            <thead>
-            <tr>
-                <th scope="col">#</th>
-                <th scope="col">Name</th>
-                <th scope="col">Surname</th>
-                <th scope="col">Email</th>
-                <th scope="col">Operation</th>
-            </tr>
-            </thead>
-            <tbody>
-            <tr v-for="user in users">
-                <th scope="row">{{ user.id }}</th>
-                <td>{{ user.name }}</td>
-                <td>{{ user.surname }}</td>
-                <td>{{ user.email }}</td>
-                <td>
-                    <button @click.prevent="open('edit', user.id )" type="button" data-bs-toggle="modal"
-                            data-bs-target="#staticBackdrop"
-                            class="btn btn-warning">Update
+    <div class="container">
+            <DataTable :value="users"
+                       :paginator="true"
+                       class="p-datatable-sm"
+                       :rows="10"
+                       responsiveLayout="scroll"
+                       paginatorTemplate="CurrentPageReport FirstPageLink PrevPageLink PageLinks NextPageLink LastPageLink RowsPerPageDropdown"
+                       :rowsPerPageOptions="[10,20]"
+                       currentPageReportTemplate="Showing {first} to {last} of {totalRecords}"
+                       scrollHeight="530px">
+                <template #header>
+                    <button type="button"  @click.prevent="open('add')" class="btn btn-success ms-2">Add</button>
+                </template>
+                <Column field="id" header="ID" sortable></Column>
+                <Column field="name" header="Name"></Column>
+                <Column field="surname" header="Surname"></Column>
+                <Column field="email" header="Email"></Column>
+                <Column field="options" header="Options">
+                    <template #body="{data}">
+                        <Button  icon="pi pi-pencil" class="p-button-rounded p-button-warning" @click.prevent="open('edit', data.id )"/>
+                        <Button  icon="pi pi-trash" class="p-button-rounded p-button-danger" @click.prevent="deleteUser($event, data.id)" />
+                    </template>
+
+                </Column>
+            </DataTable>
+            <Dialog header="Header"
+                    :modal="true"
+                    :closable="false"
+                    :visible.sync="display"
+                    :contentStyle="{
+                    width: '30vw',
+                    minHeight: '300px',
+                 }">
+                <template #header>
+                    <h5 class="modal-title" v-show="!editmode">Create User</h5>
+                    <h5 class="modal-title" v-show="editmode">Edit User</h5>
+                </template>
+
+                <form @submit.prevent="editmode ? updateUser() : createUser()" class="">
+                    <div class="form-outline mb-4">
+                        <label class="form-label">Name</label>
+                        <input v-model="form.name" type="text" class="form-control"/>
+                        <ValidationError :if="errors.hasOwnProperty('name')" :errors="errors?.name || []"/>
+                    </div>
+
+                    <div class="form-outline mb-4">
+                        <label class="form-label">Surname</label>
+                        <input v-model="form.surname" type="text" class="form-control"/>
+                        <ValidationError :if="errors.hasOwnProperty('surname')" :errors="errors?.surname || []"/>
+                    </div>
+
+                    <div class="form-outline mb-4">
+                        <label class="form-label">Email</label>
+                        <input v-model="form.email" type="email" class="form-control"/>
+                        <ValidationError :if="errors.hasOwnProperty('email')" :errors="errors?.email || []"/>
+                    </div>
+
+                    <div class="form-outline mb-4">
+                        <label class="form-label">Password</label>
+                        <input v-model="form.password" type="password" class="form-control"/>
+                        <ValidationError :if="errors.hasOwnProperty('password')" :errors="errors?.password || []"/>
+                    </div>
+
+                </form>
+
+                <template #footer>
+                    <button type="button" class="btn btn-secondary" @click.prevent="close">Close</button>
+                    <button v-show="!editmode" @click.prevent="createUser" type="button" class="btn btn-primary">
+                        Add user
                     </button>
-                    <button @click.prevent="deleteUser(user.id)" type="button" class="btn btn-danger">Delete</button>
-                </td>
-            </tr>
-            </tbody>
-        </table>
+                    <button v-show="editmode" @click.prevent="updateUser(form.id)" type="button" class="btn btn-primary">
+                        Update user
+                    </button>
+                </template>
+            </Dialog>
+            <ConfirmPopup></ConfirmPopup>
 
-        <!--    modal-->
-        <div class="modal fade" id="staticBackdrop" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1"
-             aria-labelledby="staticBackdropLabel" aria-hidden="true">
-            <div class="modal-dialog">
-                <div class="modal-content">
-                    <div class="modal-header">
-                        <h5 class="modal-title" v-show="!editmode">Create User</h5>
-                        <h5 class="modal-title" v-show="editmode">Edit User</h5>
-                        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-                    </div>
-                    <div class="modal-body">
-                        <form @submit.prevent="editmode ? updateUser() : createUser()" class="p-5">
-                            <div class="form-outline mb-4">
-                                <label class="form-label">Name</label>
-                                <input v-model="form.name" type="text" class="form-control"/>
-                                <ValidationError :if="errors.hasOwnProperty('name')" :errors="errors?.name || []" />
-                            </div>
-
-                            <div class="form-outline mb-4">
-                                <label class="form-label">Surname</label>
-                                <input v-model="form.surname" type="text" class="form-control"/>
-                                <ValidationError :if="errors.hasOwnProperty('surname')" :errors="errors?.surname || []" />
-                            </div>
-
-                            <div class="form-outline mb-4">
-                                <label class="form-label">Email</label>
-                                <input v-model="form.email" type="email" class="form-control"/>
-                                <ValidationError :if="errors.hasOwnProperty('email')" :errors="errors?.email || []" />
-                            </div>
-
-                            <div class="form-outline mb-4">
-                                <label class="form-label">Password</label>
-                                <input v-model="form.password" type="password" class="form-control"/>
-                                <ValidationError :if="errors.hasOwnProperty('password')" :errors="errors?.password || []" />
-                            </div>
-
-                        </form>
-                    </div>
-
-                    <div class="modal-footer">
-                        <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-                        <button v-show="!editmode" @click.prevent="createUser" type="button" class="btn btn-primary">Add
-                            user
-                        </button>
-                        <button v-show="editmode" @click.prevent="updateUser(form.id)" type="button"
-                                class="btn btn-primary">Update user
-                        </button>
-                    </div>
-                </div>
-            </div>
-        </div>
     </div>
+
 </template>
 
 <script>
 import Form from 'vform'
 import ValidationError from "./errors/ValidationError";
+import DataTable from 'primevue/datatable';
+import Column from 'primevue/column';
+import ColumnGroup from 'primevue/columngroup';
+import Dialog from 'primevue/dialog';
+import Button from 'primevue/button'
+import ConfirmPopup from 'primevue/confirmpopup';
+
 
 export default {
     name: "Users",
+    components: {
+        DataTable,
+        Column,
+        ColumnGroup,
+        ValidationError,
+        Dialog,
+        ConfirmPopup,
+        Button
+    },
 
     data() {
         return {
+            display: false,
             users: [],
             errors: [],
             isShowModal: false,
@@ -107,9 +119,6 @@ export default {
 
             })
         }
-    },
-    components: {
-        ValidationError
     },
     mounted() {
         this.getUsers()
@@ -131,9 +140,8 @@ export default {
 
                         return;
                     }
-                    const url = new URL('/users', window.location.origin)
-                    window.location.href = url.toString()
-                    console.log('success', r.data);
+                    this.display = false
+                    this.getUsers()
                 });
         },
 
@@ -149,23 +157,36 @@ export default {
 
                         return;
                     }
-                    const url = new URL('/users', window.location.origin)
-                    window.location.href = url.toString()
-                    console.log('success', r.data);
+                    this.display =false
+                    this.getUsers()
                 });
 
         },
-        deleteUser(id) {
-            axios.delete(`/api/user/${id}`)
-                .then(res => {
-                    this.getUsers()
-                })
+        deleteUser(event, id) {
+
+            this.$confirm.require({
+                target: event.currentTarget,
+                message: 'Do you want to delete this record?',
+                icon: 'pi pi-info-circle',
+                acceptClass: 'p-button-danger',
+                accept: () => {
+                    axios.delete(`/api/user/${id}`)
+                        .then(res => {
+                            this.getUsers()
+                        })
+                },
+                reject: () => {
+                    this.$confirm.close();
+                }
+            });
+
         },
 
         open(mode, id = null) {
             if (mode === 'add') {
                 this.form.reset()
                 this.editmode = false
+
             }
             if (mode === 'edit') {
                 this.editmode = true
@@ -173,11 +194,25 @@ export default {
                 this.form.fill(user)
             }
 
+            this.display = true
+
+        },
+        hide(){
+            this.display = false
+            this.form.reset()
+            this.errors = []
+        },
+        close() {
+            this.display = false
+            this.form.reset()
+            this.errors = []
         }
     }
 }
 </script>
 
 <style scoped>
+.custom-dialog{
 
+}
 </style>
